Added game cursor, build menu UI UI and cursor info UI
In the game scene a new cursor has been added that allows the player to move around the game.
Using mset() info about the tile the cursor is on is also displayed to the player as well as text ready for pop count and happiness.
Build menu UI has also been added with cursor and options for buildings.